Issue #2: Have two topics - Sample and Trigonometry
Introducing class ly_group to represent one group.  The group can either
contain a "question bank" or "list of questions" or it can contain
subgroups.

The global variable ly_groups is an array of ly_group objects.  When the
app is loaded we show a list of top level groups.  When a group is
selected, we either show sub-groups or if the group is a "bunch of
questions" then start display of MCQs.

Introducing QSet abstract base class.  A QSet is a "set of questions" or
"bunch of questions" or "a practice" or "group of questions".  Adding
SampleQSet and Trigonometry derived classes of QSet.